Ye Olde Brick, hacked one very late night in 2010.
Full of inefficient or maybe broken code, as well
as the occasional memory leak. I am not proud of
it and I was *really* tired back then. While this
was an experiment in writing (f)ugly code and I
just wanted something running, I have to admit
that this code has served me well ever since.

Too well to be bothered to ever look at it again,
really...